requirements.txt
# RAGline Worker Service Dependencies

# Core Celery
celery[redis]==5.3.4
redis==5.0.1
kombu==5.3.4

# Async pools
gevent==23.9.1
eventlet==0.33.3

# System monitoring
psutil==5.9.6

# HTTP and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.1

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Environment and config
python-dotenv==1.0.0

# LLM and RAG
openai==1.6.1
tiktoken==0.5.2
sse-starlette==1.8.2
pgvector==0.2.4

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
